Initial commit of project files
Adding firmware and board designs. The firmware is a known good build
that would be suitable for a basic GPIO test of the controller. It could
be ported to others easily, but the pin definition is for the
CM-164K64-A version. I'll add the pin and clock Kinetis tool files in a
separate directory to include other configurations.

The project is built from the ground up using KDS SDK V3.2.0 generated
by mcuexpresso.

KiCAD designs are for two different boards. One is a IO module for the
system, includes six 4A MOSFETs to drive LEDs or other things. If you
use this design to drive motors or anything with a large inductive load,
please add a suitable flyback diode to keep from blowing the drives out.
This doesn't really have a full BOM included, I'll update that in the
future.

The second board is a strip for mounting LED strips and sharing a drive
input.

*Disclaimer* This project is mostly to kick the tires of the system and
is not optomized for cost or for using more typical solutions for doing
the same thing. It is possible that there are other ways to do this same
end result for a lot less or using less high performance MCUs, but at
the same time, doing smooth color blending and interactions for 48
different outputs and then having enough input left over to do
interesting interface is a pretty cool thing in my book.