Addding copy of existing firmware as a separate test firmware
Just getting this stuff sorted on how to organize it best. I'm copying
the existing firmware over to a test folder to keep it for a basic test
app, so it is easy to maintain. Initially I was thinking there is
separate branches or whatever, but that doesn't work well since you end
up having to sync to a particular branch to work on it. This way it is
all in one repo. Not sure if all projects will be set up this way or if
we will do a new repo for each one, but I think it makes sense to do the
firmware (app), firmware-test, and the hardware to support all of it.